fix: cast cost_center_count to float8 for consistent data type in selectLedger query

+++ src/db/others/query/accounts.js
@@ -70,7 +70,7 @@ export async function selectLedger(req, res, next) {
 			value: accountSchema.ledger.uuid,
 			label: accountSchema.ledger.name,
 			has_cost_center: sql`CASE WHEN cost_center.cost_center_count > 0 THEN TRUE ELSE FALSE END`,
-			cost_center_count: sql`COALESCE(cost_center.cost_center_count, 0)`,
+			cost_center_count: sql`COALESCE(cost_center.cost_center_count, 0)::float8`,
 		})
 		.from(accountSchema.ledger)
 		.leftJoin(
